LG2
git-svn-id: svn+ssh://beta.visl.sdu.dk/usr/local/svn/repos/visl/external/LearnGreenlandic2@7928 cb2587b3-b7ff-0310-8c81-a2a651690ada

+++ src/LearnGreenlandic2.cpp
@@ -100,7 +100,7 @@ int main(int argc, char *argv[]) {
         settings.beginWriteArray("paths", z);
         for (dirmap_t::const_iterator it = dirs.begin() ; it != dirs.end() ; ++it, ++i) {
             settings.setArrayIndex(i);
-            settings.setValue("revision", it->first);
+            settings.setValue("revision", uint(it->first));
             settings.setValue("path", it->second);
         }
         settings.endArray();
